Refactor code into more files
Each struct gets its own file
