+++ cmd/utils.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"fmt"
-	"github.com/google/gopacket/pcap"
 	log "github.com/sirupsen/logrus"
 	"strings"
 	"time"
@@ -26,45 +25,6 @@ func stringPrefixInSlice(a string, list []string) bool {
 		}
 	}
 	return false
-}
-
-// Uses libpcap to get a list of configured interfaces
-// and populate the Interfaces.
-func getConfiguredInterfaces() {
-	if len(Interfaces) > 0 {
-		return
-	}
-	ifs, err := pcap.FindAllDevs()
-	if err != nil {
-		log.Fatal(err)
-	}
-	for _, i := range ifs {
-		if len(i.Addresses) == 0 {
-			continue
-		}
-		Interfaces[i.Name] = i
-	}
-}
-
-// Print out a list of all the interfaces that libpcap sees
-func listInterfaces() {
-	getConfiguredInterfaces()
-	for k, v := range Interfaces {
-		fmt.Printf("Interface: %s\n", k)
-		for _, a := range v.Addresses {
-			ones, _ := a.Netmask.Size()
-			if a.Broadaddr != nil {
-				fmt.Printf("\t- IP: %s/%d  Broadaddr: %s\n",
-					a.IP.String(), ones, a.Broadaddr.String())
-			} else if a.P2P != nil {
-				fmt.Printf("\t- IP: %s/%d  PointToPoint: %s\n",
-					a.IP.String(), ones, a.P2P.String())
-			} else {
-				fmt.Printf("\t- IP: %s/%d\n", a.IP.String(), ones)
-			}
-		}
-		fmt.Printf("\n")
-	}
 }
 
 // takes a list of ports and builds our BPF filter
